### 2025-08-03 #718 @cforgaci
rcrisp:  Automate the Delineation of Urban River Spaces
Submitting Author Name: Claudiu Forgaci
Submitting Author Github Handle: @cforgaci
Repository:  https://github.com/CityRiverSpaces/rcrisp
Version submitted:
Submission type: Stats
Badge grade: gold

Version accepted: TBD
---

```
Package: rcrisp
Title: Automate the Delineation of Urban River Spaces
Version: 0.1.4
    person("Claudiu", "Forgaci", , "c.forgaci@tudelft.nl", role = c("aut", "cre", "cph"),
    person("Francesco", "Nattino", , "f.nattino@esciencecenter.nl", role = "aut",
    person("Fakhereh", "Alidoost", , "f.alidoost@esciencecenter.nl", role = "ctb",
    person("Meiert Willem", "Grootes", , "m.grootes@esciencecenter.nl", role = "ctb",
    person("Netherlands eScience Center", , , "info@esciencecenter.nl", role = "fnd")
  )
Description: Provides tools to automate the morphological delineation of
    riverside urban areas based on a method introduced in Forgaci (2018)
    <doi:10.7480/abe.2018.31>. Delineation entails the identification of
    corridor boundaries, segmentation of the corridor, and delineation of
    the river space using two-dimensional spatial information from street
    network data and digital elevation data in a projected CRS. The
    resulting delineation can be used to characterise spatial phenomena
    that can be related to the river as a central element.
License: Apache License (>= 2)
    checkmate,
    dbscan,
    lwgeom,
    osmdata,
    rcoins,
    rstac,
    sf (>= 0.9.0),
    sfheaders,
    sfnetworks,
    terra,
    tidygraph,
    units,
    visor
    withr
    knitr
Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [x] A pre-submission inquiry has been approved in [issue#716](#716)

The target audience are researchers and practitioners concerned with the analysis of riverside urban phnomena. As a tool for automatic spatial delineation, `rcrisp` tackles the Modifiable Areal Unit Problem by providing _an alternative to arbitrary spatial units used in current practice_ (e.g., circular, rectangular or polygonal "cut-outs", economically or politically driven development boundaries, country-specific census boundaries), namely _spatial-morphological units_ derived from both the topography of the river valley and the configuration of the urban fabric. As a tool for _automated_ river space delineation, `rcrisp` enables (1) _local integrated analysis_ because it provices neutral boundaries that can be used to reliably integrate various data sources, and (2) _global cross-case comparative analyses_ as the resulting delineations are neither too context-specific (e.g, a country-specific census tracts) nor too out-of-context (e.g., arbitrary cut-outs) than other delineation approaches.

`rcrisp` is the first implementation of a novel algorithm.

N/A

gold

`rcrisp` fulfils at least three of the four aspects, as follows:

1. We are confident that we have identified most (if not all) standards that are currently applicable and we consider the current level of compliance to be well above minimal. We look forward to determining together with the reviewer(s) the potential level of compliance and will do our best to reach it.

2. `rcrisp` currently demonstrates extensive compliance with sub-categories G1, G2 and G5 and it meets most standards in sub-categories SP1, SP2, and SP4. It also complies with at least two standards in each of G4, SP3 and SP6.

3. As a package meant to define analytical boundaries, `rcrisp` presents broad applicability in any kind of analysis concerned with riverside phenomena that can benefit from morphological boundaries. This also means that the package is not bound to any specific discipline: it can be used in ecology, hydrology, urban studies, transport planning, to name a few disciplines. In terms of scale, `rcrisp` enables use cases of local analyses (e.g., delineation of a specific riverspace, segment or corridor of interest in a city) as well as cross-case analyses (e.g., delineation of any of those spatial units for a large number of river-crossed cities for purposes on comparison a large-scale understanding of a phenomenon of interest). Locally it enables *integration* of various data sources within a single neutral boundary, while globally it allows for *scalability*.

4. The current design and API of `rcrisp` is structured around the main user-facing `delineate_*` functions. In addition to modules dedicated to each of those delineation tasks, we designed specific modules for delineation data retrieval (currently limited to OSM and GLO-30 DEM data), example data retrieval, spatial network data pre-processing, and utility functions. This design is the result of our efforts to make the package both easily maintainable and understandable to users. While this design has been effective for our purposes, we look forward to any suggestions on how we can improve it further.

As mentioned in the [presubmission issue](https://github.com/ropensci/software-review/issues/716#issuecomment-3144247108),

> We [...] did our best to address the issues flagged by `autotest`, discarding those we considered to be false positives. However, `test-autotest` only passes when run interactively (see a detailed description of this issue in https://github.com/CityRiverSpaces/rcrisp/pull/321#discussion_r2247242258).
