### 2025-08-04 #718 @mpadge
Thanks @cforgaci , very happy to see this reach the full-submission stage! We'll start trying to find a handling editor as soon as we can, but please note that several of our editors will be away for the next couple of weeks, so this may take some time. You'll be notified here as soon as we've found somebody.
