### 2025-08-19 d7c70b8
Add metadata change to changelog
