+++ cran-comments.md
@@ -1,17 +1,14 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
-
-Days since last update: 2
+0 errors | 0 warnings | 0 note
 
 ## Resubmission
 
-This is a resubmission in response to a problem pointed out on CRAN.
-In this version:
+* To fix the current problems on CRAN: `get_osm_example_data()` and
+`get_dem_example_data()` were updated to fail gracefully in examples, vignettes
+and tests, that is, return a message and NULL, if internet resource is not
+available.
 
-* `check_cache()` is moved from `.onLoad()` to `.onAttach()` to comply with
-  CRAN policy on namespace loading.
-* The NOTE on namespace load is resolved.
-* In addition, the CRAN badge is added to the README, the DOI from Zenodo is
-  added to CITATION.cff, DESCRIPTION and README (badge), and vignette file
-  names are updated to ensure they are listed in order on CRAN.
+* In addition, this version includes extensive documentation updates, adds input
+assertions throughout the package, removes an unused function, and replaces
+`sapply()` with `vapply()` throughout the package.
