### 2025-08-19 90adc9a
run flatten_network examples in interactive mode
