### 2025-08-19 73644a8
Do not run vignette codechunks with no exampledata
