+++ vignettes/vig_03-network-preparation.Rmd
@@ -22,6 +22,12 @@ library(sfnetworks)
 
 bucharest_osm <- get_osm_example_data()
 bucharest_dem <- get_dem_example_data()
+
+if (any(is.null(bucharest_osm), is.null(bucharest_dem))) {
+  cat("NOTE: Example data was not found; ",
+      "subsequent code chunks will be skipped.\n", sep = "")
+  knitr::opts_chunk$set(eval = FALSE)
+}
 ```
 
 ```{r srr-tags, eval=FALSE, echo=FALSE}
