+++ vignettes/vig_04-valley-delineation.Rmd
@@ -50,7 +50,7 @@ bucharest_dem <- get_dem_example_data()
 
 if (any(is.null(bucharest_osm), is.null(bucharest_dem))) {
   cat("NOTE: Example data was not found; ",
-    "subsequent code chunks will be skipped.\n", sep = "")
+      "subsequent code chunks will be skipped.\n", sep = "")
   knitr::opts_chunk$set(eval = FALSE)
 }
 
