### 2025-08-20 d864e5a
Disable unnecessary eval of delineate code chunk
