+++ vignettes/vig_05-corridor-delineation.Rmd
@@ -48,14 +48,16 @@ We start by demonstrating the use of the all-in-one `delineate()` function which
 
 The `delineate()` function carries out the above steps in one and returns a list that by default contains the following output elements: `corridor`, `segments`, and `riverspace`.
 
-```{r delineate, fig.alt="Delineation of the corridor of River Dâmbovița in Bucharest"}
+```{r delineate, eval=FALSE}
 bucharest_dambovita <- delineate(
   city_name,
   river_name,
   segments = TRUE,
   riverspace = TRUE
 )
+```
 
+```{r plot, fig.alt="Delineation of the corridor of River Dâmbovița in Bucharest"}
 # Plot all layers within the extent of the delineated corridor
 bbox <- st_bbox(bucharest_dambovita$corridor)
 plot(bucharest_dambovita$valley, col = "grey", border = NA,
