+++ cran-comments.md
@@ -4,10 +4,10 @@
 
 ## Resubmission
 
-* To fix the current problems on CRAN: `get_osm_example_data()` and
-`get_dem_example_data()` were updated to fail gracefully in examples, vignettes
-and tests, that is, return a message and NULL, if internet resource is not
-available.
+* This is a resubmission to solve problems identified by package checks on CRAN.
+Specifically, we have updated `get_osm_example_data()` and
+`get_dem_example_data()` to fail gracefully in examples, vignettes and tests,
+that is, return a message and NULL, if the internet resource is not available.
 
 * In addition, this version includes extensive documentation updates, adds input
 assertions throughout the package, removes an unused function, and replaces
