### 2025-08-21 ed5c9a8
Remove periods from titles in function docs
