+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   vig_05-corridor-delineation: vig_05-corridor-delineation.html
   vig_06-corridor-segmentation: vig_06-corridor-segmentation.html
   vig_07-riverspace-delineation: vig_07-riverspace-delineation.html
-last_built: 2025-08-21T09:50Z
+last_built: 2025-08-21T10:13Z
 urls:
   reference: https://cityriverspaces.github.io/rcrisp/reference
   article: https://cityriverspaces.github.io/rcrisp/articles
