### 2025-08-29 #718 @mpadge
@cforgaci I've decided to handle this one myself. A general statement for the record: The input data for this package come from my own package, https://github.com/ropensci/osmdata. While that is potentially a conflict of interest, I'm genuinely interested in learning how that package is used by others. Also ping @jmaspons, who has recently become the primary maintainer of osmdata.
