### 2025-08-29 #718 @mpadge
---

- The package seems very well developed, and the documentation is extensive and generally excellent. One thing I would nevertheless have liked to see is a general overview of a typical workflow. The README provides an example of "how to solve a common problem," but gives no real indication of what that problem might be, or why it might be good to solve it. The vignettes clearly have all the pieces needed for the general workflow of the package to be used, yet there is no real demonstration of how everything can or should be used together. I suspect that the review process would be greatly aided by first providing an overview of how the various pieces of the package can and should be used within single workflows. If you decide to tackle this before we proceed to review, please do so via an issue in your repo, and link to this issue. Then we'll see your issue linked below, and will also be able to see when it has been closed.

- Your use of issues on GitHub is really good, and provides great insight into thoughts behind the package development. There are nevertheless a lot of currently open issues, many of which are pegged to [a milestone on GitHub](https://github.com/CityRiverSpaces/rcrisp/milestone/1) which has a due date of 31 Aug 2025. That suggests the package might change quite a bit in the next short while - can you provide an update on current status, stability, and likely changes before we proceed to review? More generally, it might be good to try to close as many outstanding issues as possible before proceeding to review.

- The first sentence of the README states that the tool "follow[s] the CRiSp method", yet without explaining what this is, or linking to anywhere else. It would be better to provide a brief initial explanation, and then to provide links there to both to the method vignette and the main paper. (See also first comment above.)
