### 2025-08-29 CityRiverSpaces/rcrisp#328 @cforgaci
Add a general overview of a typical workflow
As indicated in https://github.com/ropensci/software-review/issues/718#issuecomment-3236328761, we are missing:

- [ ] a general overview of a typical workflow, binding the pieces described in the vignettes together; and
- [ ] a clear indication of what the problem might be and why it might be good to solve.
