### 2025-08-30 2cf2e45
Add getting started vignette
