### 2025-08-30 4921bc3
Remove unnecessary library call
