### 2025-08-30 57c4347
Update headings and text
