### 2025-08-30 df4865d
Update description in README
