### 2025-08-30 5371e91
Update references in README and vignette
