### 2025-08-30 d5fbcf8
Update comment in README
