+++ R/osmdata.R
@@ -254,6 +254,10 @@ get_osm_city_boundary <- function(bb, city_name, crs = NULL, multiple = FALSE,
     dplyr::bind_rows(osmdata_sf$osm_polygons, osmdata_sf$osm_multipolygons) |>
       # filter using any of the "name" columns (matching different languages)
       match_osm_name(city_name_clean) |>
+      dplyr::filter(
+        suppressWarnings(as.numeric(admin_level)) ==
+          max(suppressWarnings(as.numeric(admin_level)), na.rm = TRUE)
+      ) |>
       sf::st_geometry()
   }
 
