+++ pkgdown.yml
@@ -2,6 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
+  getting-started: getting-started.html
   vig_01-method: vig_01-method.html
   vig_02-getting-osm-data: vig_02-getting-osm-data.html
   vig_03-network-preparation: vig_03-network-preparation.html
@@ -9,7 +10,7 @@ articles:
   vig_05-corridor-delineation: vig_05-corridor-delineation.html
   vig_06-corridor-segmentation: vig_06-corridor-segmentation.html
   vig_07-riverspace-delineation: vig_07-riverspace-delineation.html
-last_built: 2025-08-25T12:13Z
+last_built: 2025-09-08T21:08Z
 urls:
   reference: https://cityriverspaces.github.io/rcrisp/reference
   article: https://cityriverspaces.github.io/rcrisp/articles
