### 2025-09-08 661bedd
Prioritise exact OSM name matches (#329)
* Prioritise exact OSM name matches

* Update changelog
