### 2025-09-08 22f5c41
Get city boundary at correct admin level (#332)
* Prioritise exact OSM name matches

* Update changelog

* Do not retrieve county boundary

* Add admin level to mock city boundary

* Fix lint
