### 2025-09-09 b41856f
Update NEWS.md
Co-authored-by: Claudiu Forgaci <33600128+cforgaci@users.noreply.github.com>
