+++ NEWS.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Overall workflow was documented in a new vignette and the package README
 - Badges to Research Software Directory and R-universe were added
-- Corridors are also allowed to be MultiPolygons when delineating segments
+- Corridors are also allowed to be multipolygons when delineating segments
 
 ## Fixed
 
