### 2025-09-09 7e07b81
Specify disabled code chunk evaluation in changelog
