### 2025-09-09 b74afb3
Disable code chunk in getting started vignette (#334)
* Disable code chunk in getting started vignette

* Specify disabled code chunk evaluation in changelog
