### 2025-09-09 3c3dd88
Create separate function to get OSM river surface
