+++ NEWS.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bug returning county boundary instead of city boundary was fixed.
 - Code chunk with OSM data retrieval was disabled in getting started vignette.
 
+## Changed
+
+- OSM river surface retrieval was moved to a separate function.
+
 # rcrisp 0.2.0 - 2025-08-21
 
 ## Added
