### 2025-09-10 9c82a1b
Error when river has no crossings (#335)
