### 2025-09-11 b631f25
fix get river surface function
