+++ R/osmdata.R
@@ -165,19 +165,13 @@ get_osmdata <- function(
   river_centerline <- get_osm_river_centerline(
     bb, river_name, crs = crs, force_download = force_download
   )
-  river_surface <- get_osm_river_surface(
-    bb, river_name, crs = crs, force_download = force_download
-  )
 
-  osm_data <- list(
-    bb = bb,
-    river_centerline = river_centerline,
-    river_surface = river_surface
-  )
+  osm_data <- list(bb = bb, river_centerline = river_centerline)
 
   # Retrieve streets and railways based on the aoi
   if (!is.null(network_buffer)) {
-    aoi_network <- get_river_aoi(river, bb, buffer_distance = network_buffer)
+    aoi_network <- get_river_aoi(river_centerline, bb,
+                                 buffer_distance = network_buffer)
     osm_data <- append(osm_data, list(aoi_network = aoi_network))
     osm_data <- append(osm_data, list(
       streets = get_osm_streets(aoi_network, crs = crs,
@@ -189,14 +183,19 @@ get_osmdata <- function(
     ))
   }
 
-  # Retrieve buildings based on a different aoi
+  # Retrieve buildings and water surface based on a different aoi
   if (!is.null(buildings_buffer)) {
-    aoi_buildings <- get_river_aoi(river, bb,
+    aoi_buildings <- get_river_aoi(river_centerline, bb,
                                    buffer_distance = buildings_buffer)
     osm_data <- append(osm_data, list(aoi_buildings = aoi_buildings))
     osm_data <- c(osm_data, list(
       buildings = get_osm_buildings(aoi_buildings, crs = crs,
                                     force_download = force_download)
+    ))
+    osm_data <- append(osm_data, list(
+      river_surface = get_osm_river_surface(aoi_buildings, river_centerline,
+                                            crs = crs,
+                                            force_download = force_download)
     ))
   }
 
@@ -294,8 +293,8 @@ get_osm_city_boundary <- function(bb, city_name, crs = NULL, multiple = FALSE,
 #' @param crs Coordinate reference system as EPSG code
 #' @param force_download Download data even if cached data is available
 #'
-#' @return A list with the river centreline as object of class
-#'   [`sf::sfc_LINESTRING`] or [`sf::sfc_MULTILINESTRING`].
+#' @return The river centreline as object of class [`sf::sfc_LINESTRING`] or
+#'   [`sf::sfc_MULTILINESTRING`].
 #' @export
 #'
 #' @examplesIf interactive()
@@ -356,20 +355,21 @@ get_osm_river_centerline <- function(bb, river_name, crs = NULL,
 
 #' Get the river surface from OpenStreetMap
 #'
-#' @param bb Bounding box of class `bbox`
-#' @param river_name The name of the river as character vector of length 1,
-#'   case-sensitive.
+#' @param aoi Area of interest as sf object or bbox
+#' @param river_centerline The river centerline as an object of class
+#'   [`sf::sfc_LINESTRING`] or [`sf::sfc_MULTILINESTRING`]
 #' @param crs Coordinate reference system as EPSG code
 #' @param force_download Download data even if cached data is available
 #'
-#' @return A list with the river surface of class [`sf::sfc_POLYGON`] or
+#' @return The river surface as object of class [`sf::sfc_POLYGON`] or
 #'   [`sf::sfc_MULTIPOLYGON`].
 #' @export
 #'
 #' @examplesIf interactive()
 #' bb <- get_osm_bb("Bucharest")
 #' crs <- get_utm_zone(bb)
-#' get_osm_river_surface(bb = bb, river_name = "Dâmbovița", crs = crs,
[…]
