+++ vignettes/vig_02-getting-osm-data.Rmd
@@ -42,6 +42,7 @@ Using the obtained bounding box, we get the different layers of OSM data needed 
 city_boundary <- get_osm_city_boundary(bb, city_name, crs)
 river_centerline <- get_osm_river_centerline(bb, river_name, crs)
 river_surface <- get_osm_river_surface(bb, river_name, crs)
+river <- c(river_centerline, river_surface)
 aoi_network <- get_river_aoi(river, bb, buffer_distance = network_buffer)
 streets <- get_osm_streets(bb, crs)
 railways <- get_osm_railways(bb, crs)
