### 2025-09-11 18e1eb8
replace weird structure with append
