### 2025-09-11 a567470
Add changelog entry
