### 2025-09-12 85877e1
Update vignettes/vig_02-getting-osm-data.Rmd
Co-authored-by: Claudiu Forgaci <33600128+cforgaci@users.noreply.github.com>
