### 2025-09-12 421b519
Transform bb into sfc in `get_osm_buildings()` (#338)
* Transform bb into sfc

* Add changelog entry

* Add test

* Fix test

* Fix lint
