### 2025-09-12 52c05cb
allow for empty water surface
