+++ R/osmdata.R
@@ -396,8 +396,6 @@ get_osm_river_surface <- function(bb, river_centerline, crs = NULL,
     sf::st_filter(river, .predicate = sf::st_intersects) |>
     sf::st_union()
 
-  if (sf::st_is_empty(river_surface)) stop("No river surface geometry found.")
-
   if (!is.null(crs)) river_surface <- sf::st_transform(river_surface, crs)
 
   river_surface
@@ -557,7 +555,7 @@ get_osm_railways <- function(aoi, crs = NULL, railway_values = "rail",
 get_osm_buildings <- function(aoi, crs = NULL, force_download = FALSE) {
   # Check input
   crs <- as_crs(crs)
-  aoi <- sf::st_as_sfc(aoi)
+  aoi <- as_sfc(aoi)
 
   buildings <- osmdata_as_sf("building", "", aoi,
                              force_download = force_download)
