### 2025-09-12 faac1ee
Update tests/testthat/test-osmdata.R
Co-authored-by: Claudiu Forgaci <33600128+cforgaci@users.noreply.github.com>
