+++ NEWS.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Overall workflow was documented in a new vignette and the package README
 - Badges to Research Software Directory and R-universe were added
 - Corridors are also allowed to be multipolygons when delineating segments
+- Badge with Status at rOpenSci Software Peer Review was added
 
 ## Fixed
 
