### 2025-10-10 103aa97
Fail gracefully on stac url retrieval
