### 2025-10-12 15a1b7d
Update overall phrasing and emphasize package names
