### 2025-10-12 28eb729
Update description of core algorithms and modules
