### 2025-10-12 ff9a546
Add grant statement
