### 2025-10-12 bb7bc1f
Add method diagram
