### 2025-10-13 5f37038
Skip sstac asset url retrieval on cran
