### 2025-10-13 4c7bc11
Fix "wrong city name" test (#345)
* Suppress warning

* Update changelog

* Fix lint
