### 2025-10-13 0008174
Skip test when stac asset urls retrieval fails (#343)
* Fail gracefully on stac url retrieval

* Update changelog

* Skip stac asset url retrieval on cran
