### 2025-10-13 1a2e1d2
Update CRAN comments
