### 2025-10-13 fe0b1a5
Bump version to 0.3.0 (#346)
* Bump minor version

* Update CRAN comments
