### 2025-10-16 #718 @mpadge
(And full disclosure, @khondula also reviewed [one of my own packages](https://github.com/ropensci/software-review/issues/330), and was extremely helpful, and [got the package in way better shape](https://github.com/ropensci/software-review/issues/330#issuecomment-534942535) that it would have been without her excellently detailed review 😄 )
