### 2025-11-12 0d04d09
Disable code chunk evaluation in network vignette
