+++ NEWS.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# rcrisp 0.3.1
+
+## Fixed
+
+- Code chunk evaluation was disabled in the network preparation vignette to prevent failed http requests on CRAN.
+
 # rcrisp 0.3.0 - 2025-10-13
 
 ## Added
