### 2025-11-12 991db7e
Update cran comments
