+++ cran-comments.md
@@ -17,4 +17,4 @@ the checks on CRAN.
 is not available. However, `osmdata` is on CRAN and we believe this to be a
 temporary issue on CRAN's side. (Note that the `osmdata` check results form the
 failing flavor "r-devel-linux-x86_64-fedora-gcc" seem to be unavailable:
-https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-gcc/osmdata-00check.html
+https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-fedora-gcc/osmdata-00check.html)
