### 2025-11-21 daaa353
test on real data disabled on CRAN and CI
