+++ tests/testthat/test-riverspace.R
@@ -1,10 +1,12 @@
-riverspace_actual <-
-  delineate_riverspace(river = bucharest_osm$river_surface,
-                       occluders = bucharest_osm$buildings,
-                       ray_length = 100)
-sf::st_crs(riverspace_actual) <- 32635
-
 test_that("The riverspace of Dâmbovița within 100m is correctly returned", {
+  skip_on_ci()
+  skip_on_cran()
+  bucharest_osm <- get_osm_example_data()
+  riverspace_actual <- delineate_riverspace(
+    river = bucharest_osm$river_surface,
+    occluders = bucharest_osm$buildings,
+    ray_length = 100
+  )
   actual_surface <- sf::st_area(riverspace_actual)
   # A tolerance of 100,000 m^2 is used to account for changes in input data
   # (buildings added or removed in OSM)
