### 2025-11-21 344893c
some tests are simplified to use small artificial data
