### 2025-11-21 12ca011
use more complex artificial data where needed
