### 2025-11-24 8fde5bc
simplify geometry
