+++ NEWS.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# rcrisp 0.3.1
+# rcrisp 0.3.1 - 2025-11-24
 
 ## Changed
 
