### 2025-11-24 48f5fe7
Make package check more robust to failing data access (#348)
* Disable code chunk evaluation in network vignette

* Bump patch version

* Add changelog

* Update cran comments

* Update cran comments

* Add missing paranthesis

* Update cran-comments.md

Co-authored-by: Francesco Nattino <49899980+fnattino@users.noreply.github.com>

* fix example

all examples can now run offline

* allow for R files in the vignettes folder

* freeze vignettes

* add helper, remove exampledata and valley testdata

* test on real data disabled on CRAN and CI

* some tests are simplified to use small artificial 
[…]
