### 2026-08-19 #718 @cforgaci
I would like to kindly check in on the progress of the review. I completely understand that this takes time; I would just like to get a sense of the expected timeline.
