### 2026-08-20 #718 @maurolepore
Thanks @cforgaci for following up.

@mpadge's availability is limited for the next few days so I'll wear his handling-editor hat temporarily to try keep the ball rolling.

I see [the review by @mpadge (as reviewer) was already posted and approved above](https://github.com/ropensci/software-review/issues/718#issuecomment-5068230682) so we're only waiting for a response from @atsyplenkov.

 @atsyplenkov has been responsive so I suggest we wait a few days before we try to reach out elsewhere.
